Major changes that resulted from the split and navigation VCs bug
All of the ViewControllers except the Onboarding are now unified into
one storyboard. I might want to take them apart later. I did this to
come closer to the course patterns to help me find the problem.

To solve the cyclic references between the root VC and the
ParseUserSession for Onboarding, I created a delegate to inform when the
user has logged in.

Created a segue to show the event details that's triggerred upon
selection.